Hard-code confluent Maven repo
Post release branches build from release versions of projects, not from
SNAPSHOT versions. This requires pulling from the remote Confluent Maven
repo. The URL of this repo is configured as a variable in pom.xml as
${confluent.maven.repo}. Unfortunately, because Maven is broken in weird
and frustrating ways this cannot be used as part of the <repository>
config because variable interpolation is not supported within that tag.
The following issue was opened in 2013 and as far as I can tell has not
been addressed as of yet (I am not hopeful that it ever will)\

http://maven.40175.n5.nabble.com/Settings-properties-not-resolved-when-used-in-repository-element-td5746974.html