enable JDBC Connector to understand a null column scale
getInt returns 0 if called to retrieve a value that is actually
null. This means that the connector can never see the difference
between a column with null scale and a column with 0 scale.

This change adds a check for null, and in combination with
https://github.com/pgjdbc/pgjdbc/commit/30843e45edc1e2bac499df2d1576c2db4d3d3309
allows improved handling of scale by letting the connector
differentiate between null and 0.

Signed-off-by: crwr45 <charlie.wheelerrobinson@gmail.com>